Conflicts:
	hermes/hermes-customer/src/main/java/com/source3g/hermes/customer/service/CustomerService.java
	hermes/hermes-merchant/src/main/java/com/source3g/hermes/merchant/service/MerchantService.java
	hermes/hermes-web/src/main/webapp/jsp/merchant/accountCenter/remindSetting.jsp

+++ hermes/hermes-device/src/main/java/com/source3g/hermes/device/service/DeviceService.java
@@ -10,6 +10,7 @@ import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Service;
 
 import com.source3g.hermes.entity.Device;
+import com.source3g.hermes.entity.merchant.Merchant;
 import com.source3g.hermes.service.BaseService;
 import com.source3g.hermes.utils.Page;
 
@@ -47,7 +48,11 @@ public class DeviceService extends BaseService {
 		return page;
 	}
 
-	public void deleteById(String id) {
+	public void deleteById(String id) throws Exception {
+		Merchant merchant=mongoTemplate.findOne(new Query(Criteria.where("deviceIds").is(new ObjectId(id))), Merchant.class);
+		if(merchant!=null){
+			throw new Exception("该盒子已被绑定");
+		}
 		super.deleteById(id, Device.class);
 	}
